+++ generator/annotated2ir.py
@@ -42,7 +42,6 @@ SPECIALS = {
     "invalid": ir.Invalid,
     "column": ir.Column,
     "auto-detect-indent": ir.AutoDetectIndent,
-    "auto-detect-in-line-indent": ir.AutoDetectInLineIndent,
 }