+++ generator/ir2annotated.py
@@ -46,8 +46,6 @@ def expr_yaml(e):
         return "<column>"
     if isinstance(e, ir.AutoDetectIndent):
         return "<auto-detect-indent>"
-    if isinstance(e, ir.AutoDetectInLineIndent):
-        return "<auto-detect-in-line-indent>"
     if isinstance(e, ir.Add):
         return {"(+)": [expr_yaml(e.a), expr_yaml(e.b)]}
     if isinstance(e, ir.Sub):